Refactor packages Init class for Level 8
