+++ src/main/java/duke/exception/EmptyArgumentException.java
@@ -1,4 +1,4 @@
-package Duke.Exception;
+package duke.exception;
 
 public class EmptyArgumentException extends DukeException {
