+++ src/main/java/duke/exception/UnknownCommandException.java
@@ -1,4 +1,4 @@
-package Duke.Exception;
+package duke.exception;
 
 public class UnknownCommandException extends DukeException {
